begin refactoring the massive cfn/index module into separate files
