+++ src/calcElapsedSeconds.ts
@@ -1,0 +1,2 @@
+const calcElapsedSeconds = (since: Date) => Math.ceil((+(new Date()) - +(since)) / 1000);
+export default calcElapsedSeconds;
